rename module to better reflect its purpose
