+++ library/alloc/src/vec/in_place_collect.rs
